app/requirements.txt
tensorflow==1.6.0
pillow==5.2.0
numpy==1.15.1
flask==1.0.2
applicationinsights==0.11.6